star_print/requirements.txt
Flask==2.2.2
python-escpos==3.0.0
Pillow==9.0.0